+++ arduino/motor_driver_ATMEGA328P/motor_driver_functions.h
@@ -15,7 +15,7 @@
 // created by gpt
 // TODO manualy calculate to ensure corect table values
 #include <avr/pgmspace.h>
-extern const uint8_t sine_table[128];
+extern const float sine_table[128];
 
 
 // Global variables
